add `IsPendingCancel` method to `OrderData`
