+++ internal/consts/consts.go
@@ -18,6 +18,20 @@ const (
 )
 
 const (
+	OrderStatusNew             = "NEW"
+	OrderStatusPartiallyFilled = "PARTIALLY_FILLED"
+	OrderStatusFilled          = "FILLED"
+	OrderStatusCancelled       = "CANCELED"
+	OrderStatusPendingCancel   = "PENDING_CANCEL"
+	OrderStatusRejected        = "REJECTED"
+	OrderStatusExpired         = "EXPIRED"
+	OrderStatusUnknown         = "UNKNOWN"
+	OrderStatusUntriggered     = "UNTRIGGERED"
+	OrderStatusTriggered       = "TRIGGERED"
+	OrderStatusDeactivated     = "DEACTIVATED"
+)
+
+const (
 	CandlesInterval          = "1m"
 	ExchangeSetupConnTimeout = 3500 // ms
 	ReadTimeout              = time.Second * 5
